% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compendium_skeleton.R
\name{compendium_skeleton}
\alias{compendium_skeleton}
\title{Generate folder structures for research compendiums}
\usage{
compendium_skeleton(name = "research_compendium", path = ".", force = FALSE, format,
comments = NULL)
}
\arguments{
\item{name}{character string: the research compendium directory name. No special characters shoud be used. Default is "research_compendium".}

\item{path}{Character string containing the directory path where test (re-recorded) sound files are found.}

\item{force}{Logical controlling whether existing folders with the same name are used for setting the folder structure. The function will never overwrite existing files or folders.}

\item{format}{A character vector with the names of the folders and subfolders to be included. Take a look at `compendiums$basic` for an example.}

\item{comments}{A character string with the comments to be added to each folder in the graphical representation of the folder skeleton printed on the console.}
}
\value{
A folder skeleton for a research compendium. In addition the structure of the compendium is printed in the console.
}
\description{
\code{compendium_skeleton} generates the folder structure of a research compendium.
}
\details{
The function takes predefined folder structures to generate the directory skeleton of a research compendium.
}
\examples{
{
compendium_skeleton(name = "mycompendium", path = tempdir(), format = compendiums$basic)
}

}
\references{
{
Araya-Salas, M. (2020), *sketchy:research compendiums for data analysis in R*. R package version 1.0.0.
}
}
\author{
Marcelo Araya-Salas (\email{marcelo.araya@ucr.ac.cr})
}
