% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_ignore_files.R
\name{git_ignore_files}
\alias{git_ignore_files}
\title{Generate folder structures for research compendiums}
\usage{
git_ignore_files(name = "research_compendium", path = ".", force = FALSE, format,
comments = NULL, packrat = FALSE, git = FALSE, clone = NULL)
}
\arguments{
\item{name}{character string: the research compendium directory name. No special characters should be used. Default is "research_compendium".}

\item{path}{path to put the package directory in. Default is current directory.}

\item{force}{Logical controlling whether existing folders with the same name are used for setting the folder structure. The function will never overwrite existing files or folders.}

\item{format}{A character vector with the names of the folders and subfolders to be included. Take a look at `compendiums$basic` for an example.}

\item{comments}{A character string with the comments to be added to each folder in the graphical representation of the folder skeleton printed on the console.}

\item{packrat}{Logical to control if packrat is initialized (\code{packrat::init()}) when creating the compendium. Default is \code{FALSE}.}

\item{git}{Logical to control if a git repository is initialized (\code{git2r::init()}) when creating the compendium. Default is \code{FALSE}.}

\item{clone}{Path to a directory containing a folder structure to be cloned. Default is  \code{NULL}. If provided 'format' is ignored. Folders starting with \code{^\\.git|^\.Rproj.user|^\\.\\.Rcheck} will be ignored.}
}
\value{
A folder skeleton for a research compendium. In addition the structure of the compendium is printed in the console. If the compendium format includes a "manuscript" folder the function saves a manuscript template in Rmd format inside that folder.
}
\description{
\code{git_ignore_files} generates the folder structure of a research compendium.
}
\details{
The function takes predefined folder structures to generate the directory skeleton of a research compendium.
}
\examples{
{
data(compendiums)

git_ignore_files(name = "mycompendium", path = tempdir(), format = compendiums$basic$skeleton)
}

}
\references{
{
Araya-Salas, M., Willink, B., Arriaga, A. (2020), sketchy: research compendiums for data analysis in R. R package version 1.0.0.

Marwick, B., Boettiger, C., & Mullen, L. (2018). Packaging Data Analytical Work Reproducibly Using R (and Friends). American Statistician, 72(1), 80-88.

Alston, J., & Rick, J. (2020). A Beginners Guide to Conducting Reproducible Research.
}
}
\seealso{
\code{\link{compendiums}}, \code{\link{print_skeleton}}
}
\author{
Marcelo Araya-Salas (\email{marcelo.araya@ucr.ac.cr})
}
