% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manuscript_example-data.R
\docType{data}
\name{manuscript_template}
\alias{manuscript_template}
\title{Manuscript template}
\format{
One .Rmd file
}
\usage{
data(manuscript_template)
}
\description{
Manuscript template
}
\keyword{datasets}
