% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_wd.R
\name{open_wd}
\alias{open_wd}
\title{Open working directory}
\usage{
open_wd(path = ".", verbose = TRUE)
}
\arguments{
\item{path}{Directory path to be opened. By default it's the working directory.}

\item{verbose}{Logical to control whether the 'path' is printed in the console. Default is \code{TRUE}.}
}
\value{
Opens the working directory using the default file browser.
}
\description{
\code{open_wd} opens the working directory in the default file browser.
}
\details{
The function opens the working directory using the default file browser 
and prints the working directory in the R console. This function aims to simplify
the manipulation of files and folders in a project.
}
\examples{
{
\donttest{open_wd()}
}

}
\references{
Araya-Salas, M., & Arriaga Madrigal, A. Y. sketchy: Create Custom Research Compendiums. R package (run \code{citation("sketchy")} for the current version).
}
\seealso{
\code{\link{spot_unused_files}}
}
\author{
Marcelo Araya-Salas (\email{marcelo.araya@ucr.ac.cr})
}
\concept{data manipulation}
