% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sketchy-package.R
\docType{package}
\name{sketchy-package}
\alias{sketchy}
\alias{sketchy-package}
\title{sketchy: create custom research compendiums}
\description{
`sketchy` is intended to facilitate the use of research compendiums for data analysis in the R environment. Standard research compendiums provide a easily recognizable means for organizing digital materials, allowing  other researchers to inspect, reproduce, and build upon that research.
}
\details{
The main features of the package are:
  \itemize{
  \item Creation of (customized) folder structures, including templates for analysis reports (Rmarkdown/quarto) and manuscripts
  \item Creation of projects published as websites on GitHub Pages (format "github_site")
  \item Recording package versions with renv
  \item Simplify the inclusion of big data files with version control software and online collaborative platforms (e.g. github)
  \item Spotting/archiving unused files and checking urls in dynamic reports
  }

License: GPL (>= 2)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/maRce10/sketchy}
  \item Report bugs at \url{https://github.com/maRce10/sketchy/issues}
}

}
\author{
Marcelo Araya-Salas & Andrea Yure Arriaga Madrigal

  Maintainer: Marcelo Araya-Salas (\email{marcelo.araya@ucr.ac.cr})
}
\keyword{internal}
