% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sketchy-package.R
\docType{package}
\name{sketchy}
\alias{sketchy}
\title{sketchy: quantifying habitat-induced acoustic signal degradation}
\description{
`sketchy` is a package intended to quantify habitat-induced degradation of (animal) acoustic signals.
}
\details{
The main features of the package are:
  \itemize{
  \item Creation of (customized) folder structure
  \item Simplify the inclusion of big data files with version control software and online collaborative platforms (e.g. github)
  }

License: GPL (>= 2)
}
\author{
Marcelo Araya-Salas, Beatriz Willink & Andrea Arriaga

  Maintainer: Marcelo Araya-Salas (\email{marcelo.araya@ucr.ac.cr})
}
