% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compendiums-data.R
\docType{data}
\name{compendiums}
\alias{compendiums}
\title{List with compendium skeletons}
\format{
list
}
\usage{
data(compendiums)
}
\description{
\code{compendiums} is a list containing the format for 2 different project folder skeletons (`compendiums$basic` & `compendiums$outpu_figures`) as well as the associated comments.
}
\keyword{datasets}
