% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compendiums-data.R
\docType{data}
\name{compendiums}
\alias{compendiums}
\title{List with compendium skeletons}
\format{
list
}
\usage{
data(compendiums)
}
\description{
List with compendium skeletons
}
\keyword{datasets}
