% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_compendium.R
\name{make_compendium}
\alias{make_compendium}
\title{Generate folder structures for research compendiums}
\usage{
make_compendium(name = "research_compendium", path = ".", force = FALSE, format,
comments = NULL, packrat = FALSE)
}
\arguments{
\item{name}{character string: the research compendium directory name. No special characters should be used. Default is "research_compendium".}

\item{path}{Character string containing the directory path where test (re-recorded) sound files are found.}

\item{force}{Logical controlling whether existing folders with the same name are used for setting the folder structure. The function will never overwrite existing files or folders.}

\item{format}{A character vector with the names of the folders and subfolders to be included. Take a look at `compendiums$basic` for an example.}

\item{comments}{A character string with the comments to be added to each folder in the graphical representation of the folder skeleton printed on the console.}

\item{packrat}{Logical to control if packrat is initiallized (\code{packrat::init()}) when creating the compendium. Default is \code{FALSE}.}
}
\value{
A folder skeleton for a research compendium. In addition the structure of the compendium is printed in the console. If the compendium format includes a "manuscript" folder the function saves a manuscript template in Rmd format inside that folder.
}
\description{
\code{make_compendium} generates the folder structure of a research compendium.
}
\details{
The function takes predefined folder structures to generate the directory skeleton of a research compendium.
}
\examples{
{
data(compendiums)

make_compendium(name = "mycompendium", path = tempdir(), format = compendiums$basic$skeleton)
}

}
\references{
{
Araya-Salas, M., Willink, B., Arriaga, A. (2020), *sketchy: research compendiums for data analysis in R*. R package version 1.0.0.

Marwick, B., Boettiger, C., & Mullen, L. (2018). *Packaging Data Analytical Work Reproducibly Using R (and Friends)*. American Statistician, 72(1), 80-88.

Alston, J., & Rick, J. (2020). *A Beginners Guide to Conducting Reproducible Research*.
}
}
\author{
Marcelo Araya-Salas (\email{marcelo.araya@ucr.ac.cr})
}
