% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sketchy-package.R
\docType{package}
\name{sketchy}
\alias{sketchy}
\title{sketchy: create custom research compendiums}
\description{
`sketchy` is intended to facilitate the use of research compendiums for data analysis in the R environment. Standard research compendiums provide a easily recognizable means for organizing digital materials, allowing  other researchers to inspect, reproduce, and build upon that research.
}
\details{
The main features of the package are:
  \itemize{
  \item Creation of (customized) folder structure
  \item Simplify the inclusion of big data files with version control software and online collaborative platforms (e.g. github)
  }

License: GPL (>= 2)
}
\author{
Marcelo Araya-Salas, Beatriz Willink & Andrea Arriaga

  Maintainer: Marcelo Araya-Salas (\email{marcelo.araya@ucr.ac.cr})
}
